pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "depo-stock"
version = "1.0.0"
description = "Cliente Python para el WebService wgesStockDepositosFiscales de ARCA/AFIP"
requires-python = ">=3.10"
license = {text = "MIT"}
dependencies = [
    "zeep>=4.2",
    "pydantic>=2.0",
    "cryptography>=41.0",
    "fastapi>=0.110",
    "uvicorn[standard]>=0.27",
    "python-dotenv>=1.0",
    "lxml>=5.0",
    "requests>=2.31",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0",
    "pytest-asyncio>=0.23",
    "httpx>=0.27",
]

[tool.hatch.build.targets.wheel]
packages = ["src/depo_stock"]

[tool.pytest.ini_options]
testpaths = ["tests"]
